General housekeeping: Make plugin buffer functions take size_t * instead of int * to match the parameter type of the buffer functions called in the core. Get rid of unsafe int * <==> size_t * casting. Use ssize_t where int was used and size_t where unsigned int was used in the buffer calls to not alter signedness in the plugins. No API version change since it should only be an issue for 64-bit sim builds.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@13233 a1c6a512-1295-4272-9138-f99709370657

+++ apps/codecs.c
@@ -63,7 +63,7 @@ void sim_codec_close(void *pd);
 extern unsigned char codecbuf[];
 #endif
 
-extern void* plugin_get_audio_buffer(int *buffer_size);
+extern void* plugin_get_audio_buffer(size_t *buffer_size);
 
 struct codec_api ci_voice;
